+++ src/ADO.Net.Client.Core/DbParameterFormatter.cs
@@ -214,7 +214,6 @@ namespace ADO.Net.Client.Core
         /// <param name="parameter"></param>
         /// <param name="parameterValue">The value of the parameter</param>
         /// <param name="info">The information.</param>
-        /// <returns></returns>
         public void MapDbParameter(IDbDataParameter parameter, object parameterValue, PropertyInfo info)
         {
             parameter.ParameterName = string.Concat(ParameterNamePrefix, info.Name);